+++ nnet/trainer.h
@@ -7,10 +7,10 @@
 using Math::Vec;
 
 namespace ML {
-  typedef pair<Vec<f64>, Vec<f64>> vec_pair;
+  typedef pair<Vec<u64>, Vec<u64>> vec_pair;
   
   struct Training_Data {
-    f64              *data;
+    u64              *data;
     vector<vec_pair> v;
     uint             pair_count;
     uint             in_vec_size;
@@ -28,7 +28,7 @@ namespace ML {
     Training_Data  *t_data;
     vector<Nnet>   netv;
     Nnet_Structure structure;
-    vector<f64>    costv;
+    vector<u64>    costv;
 
     Trainer(Nnet_Structure _structure, uint net_count);
     Trainer(Nnet &net, uint net_count);
@@ -38,10 +38,10 @@ namespace ML {
     void load_training_data(string &path);
     void save_training_data(string &path);
 
-    vector<f64> *calc_costv(uint block_size, uint rounds);
+    vector<u64> *calc_costv(uint block_size, uint rounds);
     Nnet *get_best();
     void dup_best();
 
-    vector<f64> *train(uint block_size, uint rounds, f64 t_factor);
+    vector<u64> *train(uint block_size, uint rounds, f64 t_factor);
   };
 }
